Revert "update jsoncpp to support int64 and uint64 (json2pb, pb2json need it since protobuf support int64)"
This reverts commit 70c62b23f883db3f36ea943fdd614b6a06245c27.

+++ jsoncpp-0.5.0/include/json/forwards.h
@@ -3,8 +3,6 @@
 
 # include "thirdparty/jsoncpp-0.5.0/include/json/config.h"
 
-#include <stdint.h>
-
 namespace Json {
 
    // writer.h
@@ -18,8 +16,8 @@ namespace Json {
    class Features;
 
    // value.h
-   typedef int64_t Int;
-   typedef uint64_t UInt;
+   typedef int Int;
+   typedef unsigned int UInt;
    class StaticString;
    class Path;
    class PathArgument;
